+++ core-bundle/src/Resources/contao/dca/tl_article.php
@@ -201,7 +201,7 @@ $GLOBALS['TL_DCA']['tl_article'] = array
 			(
 				array('tl_article', 'generateAlias')
 			),
-			'sql'                     => "varchar(128) COLLATE utf8_bin NOT NULL default ''"
+			'sql'                     => "varchar(128) COLLATE utf8mb4_bin NOT NULL default ''"
 		),
 		'author' => array
 		(