+++ core-bundle/src/Resources/contao/dca/tl_form.php
@@ -144,7 +144,7 @@ $GLOBALS['TL_DCA']['tl_form'] = array
 			(
 				array('tl_form', 'generateAlias')
 			),
-			'sql'                     => "varchar(128) COLLATE utf8_bin NOT NULL default ''"
+			'sql'                     => "varchar(128) COLLATE utf8mb4_bin NOT NULL default ''"
 		),
 		'jumpTo' => array
 		(