+++ core-bundle/src/Resources/contao/dca/tl_user.php
@@ -149,7 +149,7 @@ $GLOBALS['TL_DCA']['tl_user'] = array
 			'flag'                    => 1,
 			'inputType'               => 'text',
 			'eval'                    => array('mandatory'=>true, 'rgxp'=>'extnd', 'nospace'=>true, 'unique'=>true, 'maxlength'=>64, 'tl_class'=>'w50'),
-			'sql'                     => "varchar(64) COLLATE utf8_bin NULL"
+			'sql'                     => "varchar(64) COLLATE utf8mb4_bin NULL"
 		),
 		'name' => array
 		(